feat: step 1 renaming folders
