+++ src/modules/departments/mutation/departments.ts
@@ -5,7 +5,7 @@ import {
     Department,
     UpdateDepartmentInput,
 } from "cv-graphql";
-import { getQuery, createMutation, deleteMutation, updateMutation } from "../api/departments";
+import { getQuery, createMutation, deleteMutation, updateMutation } from "../apis/departments";
 
 interface DepartmentsResult {
     departments: Department[];
